requirements.txt
click
SPARQLWrapper
bs4
lxml
joblib
tqdm
requests
burst_detection
click-log
numpy
matplotlib
pandas
seaborn
scipy
sklearn
openpyxl
pdfminer.six
pymongo
mpld3
scikit-learn